+++ recipes/chef_server_config.rb
@@ -16,12 +16,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-if node['rundeck']['secret_file'].nil?
-  rundeck_secure = data_bag_item(node['rundeck']['rundeck_databag'], node['rundeck']['rundeck_databag_secure'])
-else
-  rundeck_secret = Chef::EncryptedDataBagItem.load_secret(node['rundeck']['secret_file'])
-  rundeck_secure = Chef::EncryptedDataBagItem.load(node['rundeck']['rundeck_databag'], node['rundeck']['rundeck_databag_secure'], rundeck_secret)
-end
+include_recipe 'rundeck::_data_bags'
 
 template '/etc/chef/rundeck.rb' do
   owner node['rundeck']['user']
@@ -33,7 +28,7 @@ template '/etc/chef/rundeck.rb' do
 end
 
 file '/etc/chef/rundeck.pem' do
-  content rundeck_secure['chef_rundeck_pem']
+  content node.run_state['rundeck']['data_bag']['secure']['chef_rundeck_pem']
   owner node['rundeck']['user']
   group node['rundeck']['group']
   mode '0400'